+++ src/StackExchange.Redis/RedisValue.cs
@@ -1223,5 +1223,27 @@ HaveString:
             leased = null;
             return default;
         }
+
+        /// <summary>
+        /// Get the digest (hash used for check-and-set/check-and-delete operations) of this value.
+        /// </summary>
+        internal ValueCondition Digest()
+        {
+            switch (Type)
+            {
+                case StorageType.Raw:
+                    return ValueCondition.CalculateDigest(_memory.Span);
+                case StorageType.Null:
+                    return ValueCondition.NotExists; // interpret === null as "not exists"
+                default:
+                    var len = GetByteCount();
+                    byte[]? oversized = null;
+                    Span<byte> buffer = len <= 128 ? stackalloc byte[128] : (oversized = ArrayPool<byte>.Shared.Rent(len));
+                    CopyTo(buffer);
+                    var digest = ValueCondition.CalculateDigest(buffer.Slice(0, len));
+                    if (oversized is not null) ArrayPool<byte>.Shared.Return(oversized);
+                    return digest;
+            }
+        }
     }
 }